refactor: rewrite to ecmascript
